+++ PalmPatcher/AssemblyAttributes.cs
@@ -1,4 +1,4 @@
 ﻿using R2API.Utils;
 
-[assembly: NetworkCompatibility(CompatibilityLevel.NoNeedForSync, VersionStrictness.DifferentModVersionsAreOk)]
+[assembly: NetworkCompatibility(CompatibilityLevel.EveryoneMustHaveMod, VersionStrictness.EveryoneNeedSameModVersion)]
 // [assembly: HG.Reflection.SearchableAttribute.OptIn]