+++ PalmPatcher/Utilities/Extensions/FileExtensions.cs
@@ -1,0 +1,13 @@
+﻿using System.Globalization;
+using System.IO;
+
+namespace PalmPatcher.Utilities.Extensions
+{
+    public static class FileExtensions
+    {
+        public static bool IsChildOf(this DirectoryInfo dir, DirectoryInfo other)
+        {
+            return dir != null && other != null && dir.FullName.StartsWith(other.FullName, true, CultureInfo.InvariantCulture);
+        }
+    }
+}