+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**-.-.- Changes:**
+
+* Fixed Palm Blast not being able to crit
+* Made Palm Blast a real projectile
+
 **1.0.0 Changes:**
 
 * First release